add serialize into json

+++ server/views.py
@@ -2,9 +2,9 @@ from .models import db, Client, Film
 import jwt
 import time
 from .wsgi import app
-from flask import request, Request, Response, redirect, url_for
+from flask import request, Request, Response
 from typing import List
-
+import json
 import enum
 
 
@@ -22,10 +22,10 @@ def jwt_token_required(f):
     def decorated_function(*args, **kwargs):
         token_state = get_token_state()
         if token_state == TokenState.Invalid:
-            r = lambda *args: "Invalid token"
+            r = lambda *args: Response("Invalid token", 401)
             return r(*args)
         elif token_state == TokenState.Expired:
-            r = lambda *args: "Invalid expired"
+            r = lambda *args: Response("Token expired", 401)
             return r(*args)
         else:
             return f(*args, **kwargs)
@@ -49,39 +49,33 @@ def get_token_state() -> TokenState:
     return TokenState.Valid if access_token.decode() == client.access_token else TokenState.Invalid
 
 
-def wrap_into_json(data: List[object]):
-    return str(data)
-
-
-def create_client(request: Request) -> str:
+def create_client(request: Request) -> int:
     login = request.form.get('login', None)
     password = request.form.get('password', None)
     email = request.form.get('email', None)
     if not login or not password or not email:
-        return "Not enught data"
+        return 400
     new_client = Client(login, email, password)
     db.session.add(new_client)
     db.session.commit()
-    return "{} created".format(login)
+    return 200
 
 
-def update_client(request: Request, user_id: int) -> str:
+def update_client(request: Request, user_id: int) -> int:
     client = Client.query.get(user_id)
+    if not client:
+        return 403
     login = request.form.get('login', None)
     password = request.form.get('password', None)
     email = request.form.get('email', None)
     if not login or not password or not email:
-        return "Not enught data"
-
-    if login:
-        client.login = login
-    if password:
-        client.password = password
-    if email:
-        client.email = email
+        return 400
+    client.login = login
+    client.password = password
+    client.email = email
     db.session.add(client)
     db.session.commit()
-    return "updated"
+    return 200
 
 
 @app.route('/api/1/users', methods=['GET', 'POST'])
@@ -89,10 +83,13 @@ def update_client(request: Request, user_id: int) -> str:
 def handle_users():
     if request.method == 'GET':
         clients = Client.query.all()
-        return wrap_into_json(clients)
+        return Response(json.dumps({'clients': [client.serialize() for client in clients]}), 200)
     if request.method == 'POST':
-        login = create_client(request)
-        return login
+        creating_status = create_client(request)
+        if creating_status == 200:
+            return Response("Client was created", 200)
+        elif creating_status == 400:
+            return Response("Not enough data", 400)
 
 
 @app.route('/api/1/users/<user_id>', methods=['GET', 'DELETE', 'PUT'])
@@ -100,15 +97,24 @@ def handle_users():
 def handle_user(user_id):
     if request.method == 'GET':
         client = Client.query.get(user_id)
-        return str(client.login)
+        if not client:
+            return Response('Client not exist', 403)
+        return Response(json.dumps(client.serialize()), 200)
     if request.method == 'DELETE':
         client = Client.query.get(user_id)
+        if not client:
+            return Response('Client not exist', 403)
         db.session.delete(client)
         db.session.commit()
-        return "deleted"
+        return Response("Client was deleted", 200)
     if request.method == 'PUT':
-        response = update_client(request, user_id)
-        return response
+        updating_status = update_client(request, user_id)
+        if updating_status == 200:
+            return Response("Client was updated", 200)
+        elif updating_status == 403:
+            return Response("Client not exist", 403)
+        elif updating_status == 400:
+            return Response("Not enough data", 400)
 
 
 def create_film(request: Request):
@@ -116,20 +122,20 @@ def create_film(request: Request):
     creation_year = request.form.get('creation_year', None)
     creation_country = request.form.get('creation_country', None)
     if not film_name or not creation_year or not creation_country:
-        return "Not enught data"
+        return 400
     new_film = Film(film_name, creation_year, creation_country)
     db.session.add(new_film)
     db.session.commit()
-    return "{} created".format(film_name)
+    return 200
 
 
-def update_film(request: Request, film_id: int):
+def update_film(request: Request, film_id: int) -> int:
     film = Film.query.get(film_id)
     film_name = request.form.get('film_name', None)
     creation_year = request.form.get('creation_year', None)
     creation_country = request.form.get('creation_country', None)
     if not film_name or not creation_year or not creation_country:
-        return "Not enught data"
+        return 400
 
     if film_name:
         film.film_name = film_name
@@ -139,7 +145,7 @@ def update_film(request: Request, film_id: int):
         film.creation_country = creation_country
     db.session.add(film)
     db.session.commit()
-    return "updated"
+    return 200
 
 
 @app.route('/api/1/films', methods=['GET', 'POST'])
@@ -147,10 +153,12 @@ def update_film(request: Request, film_id: int):
 def handle_films():
     if request.method == 'GET':
         films = Film.query.all()
-        return wrap_into_json(films)
+        return Response(json.dumps({'films': [film.serialize() for film in films]}), 200)
     if request.method == 'POST':
-        response = create_film(request)
-        return response
+        creating_status = create_film(request)
+        if creating_status == 200:
+            return Response('Film was created', 200)
+        return Response('Not enough data', 400)
 
 
 @app.route('/api/1/films/<film_id>', methods=['GET', 'DELETE', 'PUT'])
@@ -158,15 +166,21 @@ def handle_films():
 def handle_film(film_id):
     if request.method == 'GET':
         film = Film.query.get(film_id)
-        return str(film.film_name)
+        if not film:
+            return Response('Film not exist', 403)
+        return Response(json.dumps(film.serialize()), 200)
     if request.method == 'DELETE':
         film = Film.query.get(film_id)
+        if not film:
+            return Response('Film not exist', 403)
         db.session.delete(film)
         db.session.commit()
-        return "deleted"
+        return Response("Film was deleted", 200)
     if request.method == 'PUT':
-        response = update_film(request, film_id)
-        return response
+        update_status = update_film(request, film_id)
+        if update_status == 200:
+            return Response("Film was updated", 200)
+        return Response("Not enough data for update", 400)
 
 
 @app.route('/api/1/assessment', methods=['POST'])
@@ -176,20 +190,19 @@ def make_assessment():
     film_id = request.form.get('film_id', None)
     assessment_value = request.form.get('assessment_value', None)
     if not client_id or not film_id or not assessment_value:
-        return "not enough data"
+        return Response("Not enough data", 400)
 
     client = Client.query.get(client_id)
     film = Film.query.get(film_id)
-    response = client.make_assessment(assessment_value, film)
-    return response
+    client.make_assessment(assessment_value, film)
+    return Response("Assessment created", 200)
 
 
 @app.route('/api/1/assessment/<user_id>', methods=['GET'])
 @jwt_token_required
 def get_assessnent(user_id):
     client = Client.query.get(user_id)
-    return wrap_into_json(client.assessments)
-    # return wrap_into_json(Response('<Why access is denied string goes here...>', 401, {'WWW-Authenticate':'Basic realm="Login Required"'}))
+    return Response(json.dumps({'assessment': client.assessments}), 200)
 
 
 def generate_token(client: Client, for_access=False) -> str:
@@ -208,7 +221,7 @@ def login():
     login = request.form.get('login', None)
     password = request.form.get('password', None)
     if not login or not password:
-        return "not enought data"
+        return Response("Not enough data", 400)
     client = Client.query.filter_by(login=login).first()
     if client and client.verify_password(password):
         client.access_token = generate_token(client).decode()
@@ -216,12 +229,13 @@ def login():
 
         db.session.add(client)
         db.session.commit()
-        return str({
-            'access_token': client.access_token,
-            'refresh_token': client.refresh_token
-        })
+        return Response(
+            json.dumps(
+                {'access_token': client.access_token,
+                 'refresh_token': client.refresh_token}), 200
+        )
     else:
-        return 'User not exist'
+        return Response('Client not exist', 403)
 
 
 @app.route('/api/1/refresh_token')
@@ -234,6 +248,6 @@ def refresh_token():
         client.refresh_token = generate_token(client, for_access=False).decode()
         db.session.add(client)
         db.session.commit()
-        return "Refreshing success"
+        return Response("Refreshing success", 200)
     else:
-        return "Invalid refresh token"
+        return Response("Invalid refresh token", 401)
